Consolidate base write tests
