--- tests/test_core_io/test_dtype_validation.py
@@ -1,62 +1,0 @@
-from typing import TYPE_CHECKING
-
-import numpy as np
-import pytest
-
-from geff.core_io import write_arrays
-from geff.metadata._schema import GeffMetadata
-
-if TYPE_CHECKING:
-    from geff._typing import PropDictNpArray
-
-
-# -----------------------------------------------------------------------------
-# Integration tests for write_arrays
-# -----------------------------------------------------------------------------
-
-
-def _tmp_metadata():
-    """Return minimal valid GeffMetadata object for tests."""
-    return GeffMetadata(geff_version="0.0.1", directed=True)
-
-
-def test_write_arrays_rejects_disallowed_id_dtype(tmp_path) -> None:
-    """write_arrays must fail fast for node/edge ids with unsupported dtype."""
-    geff_path = tmp_path / "invalid_ids.geff"
-
-    # float16 is currently not allowed by Java Zarr
-    node_ids = np.array([1, 2, 3], dtype=np.float16)
-    edge_ids = np.array([[1, 2], [2, 3]], dtype=np.float16)
-
-    with pytest.warns(UserWarning):
-        write_arrays(
-            geff_store=geff_path,
-            node_ids=node_ids,
-            node_props=None,
-            edge_ids=edge_ids,
-            edge_props=None,
-            metadata=_tmp_metadata(),
-        )
-
-
-def test_write_arrays_rejects_disallowed_property_dtype(tmp_path) -> None:
-    """write_arrays must fail fast if any property array has an unsupported dtype."""
-    geff_path = tmp_path / "invalid_prop.geff"
-
-    # ids are fine (int32)
-    node_ids = np.array([1, 2, 3], dtype=np.int32)
-    edge_ids = np.array([[1, 2], [2, 3]], dtype=np.int32)
-
-    # property with disallowed dtype (float16)
-    bad_prop_values = np.array([0.1, 0.2, 0.3], dtype=np.float16)
-    node_props: dict[str, PropDictNpArray] = {"score": {"values": bad_prop_values, "missing": None}}
-
-    with pytest.warns(UserWarning):
-        write_arrays(
-            geff_store=geff_path,
-            node_ids=node_ids,
-            node_props=node_props,
-            edge_ids=edge_ids,
-            edge_props=None,
-            metadata=_tmp_metadata(),
-        )
